+++ packages/std/src/lib.rs
@@ -22,7 +22,7 @@ mod storage;
 mod traits;
 mod types;
 
-pub use crate::addresses::{CanonicalAddr, HumanAddr};
+pub use crate::addresses::{Addr, CanonicalAddr, HumanAddr};
 pub use crate::binary::Binary;
 pub use crate::coins::{coin, coins, has_coins, Coin};
 pub use crate::deps::{Deps, DepsMut, OwnedDeps};